+++ examples/Windows/Forms/Authenticate/App.config
