+++ examples/Windows/Forms/Authenticate/FrAnnouncement.resx
