+++ examples/Windows/Forms/Authenticate/Properties/Resources.resx
