+++ examples/Windows/Forms/Authenticate/Properties/Settings.settings
