+++ examples/Windows/Forms/Authenticate1/App.config
