+++ examples/Windows/Forms/Authenticate1/Authenticate1.csproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{EBCBAB94-6012-4FEA-98E1-F5FC29B24342}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>Announcement1</RootNamespace>
-    <AssemblyName>Announcement1</AssemblyName>
+    <RootNamespace>Authenticate1</RootNamespace>
+    <AssemblyName>Authenticate1</AssemblyName>
     <TargetFrameworkVersion>v4.7.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
