+++ examples/Windows/Forms/Authenticate1/Form1.resx
