+++ examples/Windows/Forms/Authenticate1/Properties/Resources.resx
