+++ examples/Windows/Forms/Authenticate1/Properties/Settings.settings
