+++ examples/Windows/Forms/Authenticate1/WebAuth.resx
