+++ examples/Windows/Forms/Authenticate1/packages.config
